requirements.txt
fastapi
uvicorn
python-multipart
python-docx
faster-whisper 